+++ web-ui/src/components/event-components/endurance-race/EnduranceRaceUI.tsx
@@ -119,6 +119,7 @@ export const EnduranceRaceUI = ({
                   name={record.blob.name}
                   color={record.blob.color}
                   renderFullName={!isMobile}
+                  detailsDialogVariant="event"
                 />
               </TableCell>
               {isMobile ? (